backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
psycopg2-binary==2.9.5
requests==2.28.2
python-dotenv==1.0.0
gunicorn==20.1.0
pandas==1.5.3
numpy==1.24.2
APScheduler==3.10.1
bcrypt==4.0.1
marshmallow==3.19.0
